fix: Complete fractional share support across all services
Updated remaining components that were still using int for quantities:

- src/trading/trade_history_service.py: TradeRecord.quantity/pm_quantity → float
- src/trading/automated_trading.py: _check_concentration → float quantity
- src/trading/alpaca_service.py: execute_trade() → float quantity
- integration/unified_workflow.py:
  - TradeResult.quantity → float
  - reduce_long/reduce_short use fractional math
  - Display formatting shows up to 4 decimals for fractional shares
  - Position/order printing handles fractional quantities

+++ src/trading/trade_history_service.py
@@ -29,7 +29,7 @@ class TradeRecord:
     """Complete record of a trade with full context"""
     ticker: str
     action: str
-    quantity: int
+    quantity: float  # Supports fractional shares
     entry_price: Optional[float] = None
     order_id: Optional[str] = None
     
@@ -67,7 +67,7 @@ class TradeRecord:
     
     # PM decision
     pm_action: Optional[str] = None
-    pm_quantity: Optional[int] = None
+    pm_quantity: Optional[float] = None  # Supports fractional shares
     pm_confidence: Optional[float] = None
     pm_reasoning: Optional[str] = None
     pm_stop_loss_pct: Optional[float] = None
